requirements.txt
requests
aws_requests_auth
boto3
pillow